Enabled HSpec and begun to write unit tests for the marshallers.
