+++ package.yaml
@@ -29,8 +29,8 @@ library:
   exposed-modules: GmshAPI
 
 tests:
-  test-do-some-ffi-calls:
-    main:                Main.hs
+   unit-tests:
+    main:                Spec.hs
     source-dirs:         test
     ghc-options:
     - -threaded
@@ -39,3 +39,13 @@ tests:
     dependencies:
     - haskell-gmsh
     - hspec
+
+   test-do-some-ffi-calls:
+    main:                Main.hs
+    source-dirs:         test-manual
+    ghc-options:
+    - -threaded
+    - -rtsopts
+    - -with-rtsopts=-N
+    dependencies:
+    - haskell-gmsh
