+++ test-manual/CubeSphereTest.hs
