+++ test-manual/Main.hs
@@ -3,12 +3,10 @@ module Main where
 import CubeSphereTest (cubeSphereMain)
 import CubeTest (cubeMain)
 import ViewTest (viewMain)
-import MarshallerTest (marshallerMain)
 
 main = do
    putStrLn ""
    putStrLn "Running some tests..."
---   cubeMain
---   cubeSphereMain
---   viewMain
-   marshallerMain
+   cubeMain
+   cubeSphereMain
+   viewMain
